requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
google-generativeai==0.4.1
supabase==2.4.1
gunicorn==21.2.0
requests==2.31.0
